chore(testing): rename to vitest-generator
