+++ packages/vite/index.ts
@@ -1,3 +1,3 @@
 export * from './src/utils/versions';
 export { viteConfigurationGenerator } from './src/generators/configuration/configuration';
-export { vitestGenerator } from './src/generators/vitest/vitest';
+export { vitestGenerator } from './src/generators/vitest/vitest-generator';
